+++ ios/TransactionListBaseCell.swift
@@ -9,11 +9,11 @@ import Foundation
 
 class TransactionListBaseCell : UITableViewCell {
   internal let duration = 0.1
-  internal let scaleTo: CGFloat = 0.97
   internal let hapticType = "select"
   
   var onItemPress: (Dictionary<AnyHashable, Any>) -> Void = { _ in }
   var row: Int? = nil
+  var scaleTo: CGFloat = 0.97
   
   func addShadowLayer(_ view: UIView) {
     let shadowLayer = CAShapeLayer()
