+++ ios/TransactionListViewManager.m
@@ -17,6 +17,7 @@ RCT_EXPORT_VIEW_PROPERTY(onRequestExpire, RCTBubblingEventBlock)
 RCT_EXPORT_VIEW_PROPERTY(onTransactionPress, RCTBubblingEventBlock)
 RCT_EXPORT_VIEW_PROPERTY(accountColor, UIColor)
 RCT_EXPORT_VIEW_PROPERTY(accountName, NSString)
+RCT_EXPORT_VIEW_PROPERTY(transformOrigin, CGPoint)
 RCT_EXPORT_VIEW_PROPERTY(onReceivePress, RCTBubblingEventBlock)
 RCT_EXPORT_VIEW_PROPERTY(onAvatarPress, RCTBubblingEventBlock)
 RCT_EXPORT_VIEW_PROPERTY(onCopyAddressPress, RCTBubblingEventBlock)
